@@ -1,5 +1,5 @@
 import test from 'ava';
-import lowercaseKeys from '.';
+import lowercaseKeys from './index.js';
 
 test('main', t => {
 	t.true(lowercaseKeys({FOO: true}).foo);
